acrescentado o middleware de autenticação para endpoint selecionados

+++ routes/api.php
@@ -4,7 +4,6 @@ use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClientController;
 use App\Http\Controllers\ScheduleController;
 use App\Http\Controllers\UserTypeController;
-use App\Models\UserType;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -31,27 +30,33 @@ Route::prefix('auth')->group(function() {
 
 });
 
+/**
+ * as rotas com o middleware() faz com que seja necessário passar o token gerado após o login para
+ * conseguir ter acesso a essas funcionalidades
+ */
 Route::prefix('userType')->group(function() {
 
-    Route::post('/create', [UserTypeController::class, 'create']);
+    Route::post('/create', [UserTypeController::class, 'create'])->middleware('auth:sanctum');
 
-    Route::get('/showTypes', [UserTypeController::class, 'show']);
+    Route::get('/showTypes', [UserTypeController::class, 'show'])->middleware('auth:sanctum');
 
-    Route::delete('/delete/{id}', [UserTypeController::class, 'delete']);
+    Route::delete('/delete/{id}', [UserTypeController::class, 'delete'])->middleware('auth:sanctum');
+
+    Route::put('/update/{id}', [UserTypeController::class, 'update'])->middleware('auth:sanctum');
 
 });
 
 Route::prefix('client')->group(function() {
 
-    Route::post('create', [ClientController::class, 'create']);
+    Route::post('create', [ClientController::class, 'create'])->middleware('auth:sanctum');
 
-    Route::get('show/{id}', [ClientController::class, 'show']);
+    Route::get('show/{id}', [ClientController::class, 'show'])->middleware('auth:sanctum');
 
-    Route::get('showAll', [ClientController::class, 'showAll']);
+    Route::get('showAll', [ClientController::class, 'showAll'])->middleware('auth:sanctum');
 
-    Route::put('/update/{id}', [ClientController::class, 'update']); 
+    Route::put('/update/{id}', [ClientController::class, 'update'])->middleware('auth:sanctum'); 
 
-    Route::delete('delete/{id}', [ClientController::class, 'delete']); 
+    Route::delete('delete/{id}', [ClientController::class, 'delete'])->middleware('auth:sanctum'); 
 
 });
 
@@ -60,12 +65,12 @@ Route::prefix('schedule')->group(function() {
     Route::post('/create', [ScheduleController::class, 'create']);
 
     #nessa rota eu posso procurar por um agendamento em especifico baseado no ID
-    Route::get('/show/{id}', [ScheduleController::class, 'show']);
+    Route::get('/show/{id}', [ScheduleController::class, 'show'])->middleware('auth:sanctum');
 
     #ja nessa eu posso listar todos os agendamentos disponiveis
-    Route::get('/show', [ScheduleController::class, 'showAll']);
+    Route::get('/show', [ScheduleController::class, 'showAll'])->middleware('auth:sanctum');
 
-    Route::delete('delete/{id}', [ScheduleController::class, 'delete']);
+    Route::delete('delete/{id}', [ScheduleController::class, 'delete'])->middleware('auth:sanctum');
 
 });
